change goal to FloatField
I have made changes to project model s.py and serializers.py.  I have change for project and pledge database table integerfield to floatfield.

+++ crowdfunding/projects/serializers.py
@@ -5,7 +5,7 @@ class ProjectSerializer(serializers.Serializer):
     id = serializers.ReadOnlyField()
     title = serializers.CharField(max_length=200)
     description = serializers.CharField(max_length=None)
-    goal = serializers.IntegerField()
+    goal = serializers.FloatField()
     image = serializers.URLField()
     is_open = serializers.BooleanField()
     date_created = serializers.DateTimeField(read_only=True)
